+++ integration-tests/database/index.ts
@@ -20,7 +20,6 @@ const testScenario = config.getNumber("scenario");
 
 let redshiftClusterArgs: aws.redshift.ClusterArgs | undefined;
 let dynamodbArgs: aws.dynamodb.TableArgs | undefined;
-let appScalingPolicyArgs: aws.appautoscaling.PolicyArgs | undefined;
 let rdsInstanceArgs: aws.rds.InstanceArgs | undefined;
 
 console.log(`Running test scenario #${testScenario}`);
@@ -56,11 +55,6 @@ switch (testScenario) {
             serverSideEncryption: {
                 enabled: true,
             },
-        };
-        appScalingPolicyArgs = {
-            resourceId: "test",
-            scalableDimension: "test",
-            serviceNamespace: "test",
         };
         break;
     case 4: