+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Improvements
 
 - Use consistent naming of policies.
-- Add rdsInstanceMultiAZEnabled and basic dynamodbTableAutoscalingEnabled policies.
+- Add rdsInstanceMultiAZEnabled policy.
 
 ## 0.1.0 (2019-11-26)